requirements.txt
boto3==1.34.10
botocore==1.34.10
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
jinja2==3.1.2
python-json-logger==2.0.7